requirements.txt
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.19.1
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
soupsieve==2.6
typer==0.15.2
typing_extensions==4.12.2
urllib3==2.3.0
prometheus_client==0.17.1
aioredis==2.0.1
asyncpg==0.30.0
playwright
flask 
google-generativeai==0.4.1
python-dotenv==1.0.1
faiss-cpu
sentence-transformers==2.6.1
scikit-learn==1.3.2
numpy==1.26.4
uvicorn[standard]
sse-starlette==1.6.5
pyinstaller==5.13.0
aider-install